chore: move s3 service to util package
