+++ src/main/java/com/ssafy/newjibs/member/service/MemberService.java
@@ -22,6 +22,7 @@ import com.ssafy.newjibs.member.dto.RegisterDto;
 import com.ssafy.newjibs.member.repository.MemberRepository;
 import com.ssafy.newjibs.member.util.MemberMapper;
 import com.ssafy.newjibs.member.util.SecurityUtil;
+import com.ssafy.newjibs.utility.S3Service;
 
 import lombok.RequiredArgsConstructor;
